♻️ Make Monster counters `size_t`

+++ Source/dead.cpp
@@ -31,7 +31,7 @@ void InitCorpses()
 
 	int8_t nd = 0;
 
-	for (int i = 0; i < LevelMonsterTypeCount; i++) {
+	for (size_t i = 0; i < LevelMonsterTypeCount; i++) {
 		if (mtypes[LevelMonsterTypes[i].type] != 0)
 			continue;
 
@@ -55,7 +55,7 @@ void InitCorpses()
 
 	stonendx = nd;
 
-	for (int i = 0; i < ActiveMonsterCount; i++) {
+	for (size_t i = 0; i < ActiveMonsterCount; i++) {
 		auto &monster = Monsters[ActiveMonsters[i]];
 		if (monster.uniqType != 0) {
 			InitDeadAnimationFromMonster(Corpses[nd], monster.type());
@@ -76,7 +76,7 @@ void AddCorpse(Point tilePosition, int8_t dv, Direction ddir)
 
 void SyncUniqDead()
 {
-	for (int i = 0; i < ActiveMonsterCount; i++) {
+	for (size_t i = 0; i < ActiveMonsterCount; i++) {
 		auto &monster = Monsters[ActiveMonsters[i]];
 		if (monster.uniqType == 0)
 			continue;
